fix(KEN-897): a run-on value, and a swept item [no-changelog]
No new consumer statement. Refusing a value one line cannot carry is what
main did before this branch replaced the reader, and the standing entry
already says the sweep refuses when an item it swept up has a conflict
replacing cannot settle — this makes that sentence true in a shape where
it was not.

with_name wrote invalid frontmatter for one shape it promises to refuse. A
flow collection is the only value that continues at column 0, so the
indent test read `name: [` as a value that had ended and left `foo]`
standing beside the new name. It refuses now, under the cause it already
has for a value one line cannot carry. The reader this branch deleted
computed the span and would not have done this, which is the point: a
precise reader replaced by a coarser one owes the shapes it no longer
sees. A collection that opens and closes on its line is still replaced
whole, as it always was.

The predicate is a frontmatter question rather than a rendering one and
lives there. with_name moves to rendered.rs, whose subject is already the
two renames a rendered tree takes and which called it — skill.rs renders
bytes, and holding the name rewrite as well had put two concepts in one
file at the size limit.

The sweep let an item it swept up be dropped. A tree stages its canonical
take-over and only then meets the harness link it cannot touch, so the
refusal rolls the staged work back and the item leaves nothing behind but
its dead stop — nothing to pair against, no block, and the run replaced
everything else without it. Probed before touching it: the same scope
returned Ok with four ops, the neighbour written and the blocked item
gone. The evidence is recorded where it still exists, at the rollback, so
the all-or-none rule sees the item it swept.

That makes one existing case refuse where it used to plan nothing, which
is the rule this branch shipped reaching a shape it had not reached: a
scope whose only swept item is blocked refuses, as the sibling case beside
it already asserted. Its subject is unchanged and now stronger — nothing
is planned, and nothing runs.
