+++ crates/core/src/render/skill/rendered.rs
@@ -1,13 +1,89 @@
 //! A rendered skill tree and the two renames that operate on it: the name
 //! a skill installs under, and the `.disabled` spelling a switched-off one
-//! keeps its content under.
+//! keeps its content under. The first is `with_name`, which rewrites the
+//! frontmatter entry a tool reads the name from, and a fork asks it of
+//! bytes that have no tree around them yet.
 //!
 //! Kept apart from `skill.rs`, which renders the bytes. Whatever holds a
 //! tree after that is asking one of these two questions, and both are here.
 
 use std::path::PathBuf;
 
-use super::{SKILL_FILE, with_name};
+use super::SKILL_FILE;
+
+/// The text with its frontmatter `name` set to `installed`, emitted as a
+/// YAML scalar so a value that would read as something else (`[copy]`,
+/// `gh #edited`) comes back quoted. Only that one line changes; every
+/// other line, and each line's own ending, stays as it was. A frontmatter
+/// carrying no `name` gets one as its first line, in the file's own line
+/// ending.
+///
+/// `Err` where no single line carries the name, saying which shape the
+/// file is in — "add a block", "close the one you have", "you named it
+/// twice" and "the name runs past its line" send a reader to four
+/// different edits. The validators say the same things plainly, and
+/// writing around any of them here would hide it.
+pub(crate) fn with_name(text: &str, installed: &str) -> std::result::Result<String, &'static str> {
+    // `split` tells a missing block from one whose closing marker is gone,
+    // and the two send a reader to different edits.
+    let (yaml, _) = crate::frontmatter::split(text).map_err(|problem| {
+        match problem.contains("unterminated") {
+            true => "its frontmatter block is never closed",
+            false => "it has no frontmatter",
+        }
+    })?;
+    let yaml_start = yaml.as_ptr() as usize - text.as_ptr() as usize;
+    let lines: Vec<&str> = yaml.split_inclusive('\n').collect();
+    let entry = format!("name: {}", crate::render::yaml_scalar(installed));
+    let mut at = yaml_start;
+    let mut found: Option<(usize, usize)> = None;
+    for (index, line) in lines.iter().enumerate() {
+        let start = at;
+        at += line.len();
+        // Only a top-level entry names the document. An indented line
+        // continues the one above it, and a comment names nothing.
+        if line.starts_with([' ', '\t', '#']) {
+            continue;
+        }
+        let Some((key, value)) = line.split_once(':') else {
+            continue;
+        };
+        if key.trim() != "name" {
+            continue;
+        }
+        if found.is_some() {
+            return Err("its frontmatter names it twice");
+        }
+        // Blank and comment-only lines attach to the entry without
+        // extending its value (YAML ignores them); real indented content
+        // continues the scalar onto another line, and one line cannot
+        // stand in for it.
+        let continued = lines[index + 1..]
+            .iter()
+            .map(|line| (line.starts_with([' ', '\t']), line.trim()))
+            .find(|(_, text)| !text.is_empty() && !text.starts_with('#'))
+            .is_some_and(|(indented, _)| indented);
+        // A flow collection continues at column 0, where the indent test
+        // reads it as a value that ended: `name: [` with `foo]` under it
+        // would leave the orphan standing beside the new name.
+        if continued || !crate::frontmatter::flow_closes(value.trim()) {
+            return Err("its frontmatter's `name` runs on past its own line");
+        }
+        found = Some((start, start + line.trim_end_matches(['\r', '\n']).len()));
+    }
+    let Some((from, to)) = found else {
+        let newline = match text.starts_with("---\r\n") {
+            true => "\r\n",
+            false => "\n",
+        };
+        return Ok(format!(
+            "{}{entry}{newline}{}",
+            &text[..yaml_start],
+            &text[yaml_start..]
+        ));
+    };
+    Ok(format!("{}{entry}{}", &text[..from], &text[to..]))
+}
 
 /// A rendered tree as apply materializes it: relative path, bytes.
 pub type Files = Vec<(PathBuf, Vec<u8>)>;
@@ -56,6 +132,121 @@ impl Rendered {
         let to = PathBuf::from("SKILL.md.disabled");
         for (rel, _) in self.files.iter_mut().filter(|(rel, _)| *rel == from) {
             *rel = to.clone();
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn with_name_quotes_what_would_read_as_something_else_and_keeps_every_other_byte() {
+        let cases = [
+            (
+                "---\nname: gh\n---\nBody.\n",
+                "[copy]",
+                "---\nname: \"[copy]\"\n---\nBody.\n",
+            ),
+            (
+                "---\nname: gh\n---\nBody.\n",
+                "gh #edited",
+                "---\nname: \"gh #edited\"\n---\nBody.\n",
+            ),
+            (
+                "---\nname : gh\ndescription: d\n---\n",
+                "mine",
+                "---\nname: mine\ndescription: d\n---\n",
+            ),
+            (
+                "---\r\nname: gh\r\n---\r\nBody.\r\n",
+                "mine",
+                "---\r\nname: mine\r\n---\r\nBody.\r\n",
+            ),
+            (
+                "---\nname: gh # old\n...\nBody.\n",
+                "mine",
+                "---\nname: mine\n...\nBody.\n",
+            ),
+            ("---\nname: \"gh\"\n---\n", "mine", "---\nname: mine\n---\n"),
+            // One closing on its own line is replaced whole, as always.
+            (
+                "---\nname: [a, b]\ndescription: d\n---\n",
+                "mine",
+                "---\nname: mine\ndescription: d\n---\n",
+            ),
+            (
+                "---\nname: gh\n  # note\ndescription: d\n---\n",
+                "mine",
+                "---\nname: mine\n  # note\ndescription: d\n---\n",
+            ),
+            // Only a top-level entry names the document, and a value the
+            // line carries whole is replaced whole — comment and all.
+            (
+                "---\nmeta:\n  name: inner\nname: outer\n---\n",
+                "mine",
+                "---\nmeta:\n  name: inner\nname: mine\n---\n",
+            ),
+            (
+                "---\nname: \"gh\" # package\n---\n",
+                "mine",
+                "---\nname: mine\n---\n",
+            ),
+            // No name at all: one is written as the first line, in the
+            // file's own ending. A key that merely starts with `name` is
+            // not one.
+            (
+                "---\r\ndescription: d\r\n---\r\nBody.\r\n",
+                "mine",
+                "---\r\nname: mine\r\ndescription: d\r\n---\r\nBody.\r\n",
+            ),
+            (
+                "---\nnames: many\n---\n",
+                "mine",
+                "---\nname: mine\nnames: many\n---\n",
+            ),
+        ];
+        for (text, name, want) in cases {
+            assert_eq!(with_name(text, name).as_deref(), Ok(want), "{text:?}");
+        }
+    }
+
+    /// No single line to stand in for: nothing to write a name into, two
+    /// of them, or a value running on past its own line. Each is a
+    /// refusal rather than a guess at which line meant it, and each says
+    /// which shape the file is in — the three send a reader to three
+    /// different edits.
+    #[test]
+    fn with_name_refuses_where_no_one_line_carries_the_name() {
+        for (text, problem) in [
+            ("Body.\n", "it has no frontmatter"),
+            (
+                "---\nname: gh\nBody.\n",
+                "its frontmatter block is never closed",
+            ),
+            (
+                "---\nname: a\nname: b\n---\n",
+                "its frontmatter names it twice",
+            ),
+            (
+                "---\nname: |\n  gh\n---\n",
+                "its frontmatter's `name` runs on past its own line",
+            ),
+            (
+                "---\nname: gh\n  continued\n---\n",
+                "its frontmatter's `name` runs on past its own line",
+            ),
+            // The value that continues at column 0.
+            (
+                "---\nname: [\nfoo]\n---\n",
+                "its frontmatter's `name` runs on past its own line",
+            ),
+            (
+                "---\nname: {\nfoo: bar}\n---\n",
+                "its frontmatter's `name` runs on past its own line",
+            ),
+        ] {
+            assert_eq!(with_name(text, "mine"), Err(problem), "{text:?}");
         }
     }
 }
